[수정] : store 위치 수정
